API service, get, create, save, validation, delete

+++ routes/ideas.js
@@ -46,7 +46,11 @@ router.post('/', async (req, res) => {
 //update an idea
 router.put('/:id', async(req, res) => {
     try{
-        const updatedIdea = await Idea.findByIdAndUpdate(
+        const idea = await Idea.findById(req.params.id);
+
+        //Match the usernames
+        if(idea.username === req.body.username){
+            const updatedIdea = await Idea.findByIdAndUpdate(
             req.params.id,
             {
                 $set: {
@@ -56,7 +60,12 @@ router.put('/:id', async(req, res) => {
             },
             { new: true }
         );
-        res.json({ success: true, data: updatedIdea });
+        return res.json({ success: true, data: updatedIdea });
+
+        }
+
+        //Usernames do not match
+        res.status(403).json({ success: false, error: 'You are not authorized to update this resource'});
     }catch(error){
         console.log(error);
         res.status(500).json({ success: false, error: 'Something went wrong' });
@@ -66,8 +75,18 @@ router.put('/:id', async(req, res) => {
 //delete an idea
 router.delete('/:id', async(req, res) => {
     try{
-        await Idea.findByIdAndDelete(req.params.id);
-        res.json({success: true, data: {}});
+        const idea = await Idea.findById(req.params.id);
+
+        //Match the usernames
+        if(idea.username === req.body.username){
+            await Idea.findByIdAndDelete(req.params.id);
+            return res.json({success: true, data: {}});
+        }
+
+        //Usernames do not match
+        res.status(403).json({ success: false, error: 'You are not authorized to delete this resource'});
+
+        
     }catch(error){
         console.log(error);
         res.status(500).json({ success: false, error: 'Something went wrong' });
